Break down large file.
